Move reapply events to api package
